+++ src/walker/mod.rs
@@ -1,4 +1,5 @@
 mod ast;
+mod printer;
 
 use std::{io, io::Write, path::Path};
